feat: add auth in submit task

+++ src/controller/pickedTaskController.js
@@ -12,51 +12,55 @@ function deleteImage(filePath) {
 }
 
 const submitTask = async (req, res) => {
-	const result = {};
-	const { task_id, user_id, rate } = req.body;
+	if (req.user) {
+		const result = {};
+		const { task_id, rate } = req.body;
+		const user_id = req.user.userId;
+		try {
+			if (req.file) {
+				const imgPath = req.file.path;
+				const timestamp = new Date().getTime();
+				const uniqueFilename = `${timestamp}-${req.file.originalname}`;
+				const targetPath = `src/uploads/${uniqueFilename}`;
+				fs.rename(imgPath, targetPath, (err) => {
+					if (err) {
+						console.error(err);
+						deleteImage(imgPath);
+						return res.status(500).json({ error: "Error uploading file" });
+					}
+				});
 
-	try {
-		if (req.file) {
-			const imgPath = req.file.path;
-			const timestamp = new Date().getTime();
-			const uniqueFilename = `${timestamp}-${req.file.originalname}`;
-			const targetPath = `src/uploads/${uniqueFilename}`;
-			fs.rename(imgPath, targetPath, (err) => {
-				if (err) {
-					console.error(err);
-					return res.status(500).json({ error: "Error uploading file" });
-					deleteImage(imgPath);
+				let query = `SELECT * FROM user_task WHERE task_id = ${task_id} AND user_id = ${user_id}`;
+				let queryResult = await client.query(query);
+
+				if (queryResult.rows.length > 0) {
+					result["error"] = true;
+					result["message"] = "Task already submitted";
+					deleteImage(targetPath);
+					res.status(400).json(result);
+					return;
+				} else {
+					query = `INSERT INTO user_task (task_id, user_id, rate, img) VALUES (${task_id}, ${user_id}, ${rate}, '${targetPath}') RETURNING *`;
+					queryResult = await client.query(query);
+
+					result["error"] = false;
+					result["message"] = "Task submitted";
+					result["data"] = queryResult.rows[0];
+
+					res.status(201).json(result);
 				}
-			});
-
-			let query = `SELECT * FROM user_task WHERE task_id = ${task_id} AND user_id = ${user_id}`;
-			let queryResult = await client.query(query);
-
-			if (queryResult.rows.length > 0) {
-				result["error"] = true;
-				result["message"] = "Task already submitted";
-				deleteImage(targetPath);
-				res.status(400).json(result);
-				return;
 			} else {
-				query = `INSERT INTO user_task (task_id, user_id, rate, img) VALUES (${task_id}, ${user_id}, ${rate}, '${targetPath}') RETURNING *`;
-				queryResult = await client.query(query);
+				result["error"] = true;
+				result["message"] = "Missing required fields";
 
-				result["error"] = false;
-				result["message"] = "Task submitted";
-				result["data"] = queryResult.rows[0];
-
-				res.status(201).json(result);
+				res.status(400).json(result);
 			}
-		} else {
-			result["error"] = true;
-			result["message"] = "Missing required fields";
-
-			res.status(400).json(result);
+		} catch (error) {
+			console.error(error);
+			res.status(500).json({ error: "Internal Server Error" });
 		}
-	} catch (error) {
-		console.error(error);
-		res.status(500).json({ error: "Internal Server Error" });
+	} else {
+		res.status(401).json({ error: "Unauthorized" });
 	}
 };
 module.exports = { submitTask };
